converting cart id from int to long

+++ domain/src/main/java/com/webstore/domain/repo/CartRepository.java
@@ -5,6 +5,6 @@ import com.webstore.domain.entity.Cart;
 import java.util.Optional;
 
 public interface CartRepository {
-    Optional<Cart> getCart(int id);
+    Optional<Cart> getCart(Long id);
     void saveCart(Cart cart);
 }
