+++ domain/src/main/java/com/webstore/domain/usecase/AddItemToCartUseCase.java
@@ -14,7 +14,7 @@ public class AddItemToCartUseCase {
     final ProductRepository productRepository;
 
     // decided not to reuse usecases inside one another so that if a case logic change dependant usecases remains un-affected.
-    public void invoke(int id, long skuNumber, int quantity) {
+    public void invoke(Long id, long skuNumber, int quantity) {
         Optional<Cart> cart = cartRepository.getCart(id);
         Optional<Product> p = productRepository.getProduct(skuNumber);
         p.ifPresent(product -> cart.ifPresent(cart1 -> cart1.addLineItem(product, quantity)));
